requirements.txt
mido
python-rtmidi
pynput
pycaw
dearpygui
comtypes
toml
obs-websocket-py
sounddevice
soundfile
numpy
openai
